+++ test-app/mocks/mockClients/dashboards/data/timeseriesDataHelper.js
@@ -3,6 +3,7 @@ const dayjs = require('dayjs')
 const splitIntoRandomValues = (total, parts) => {
   let arr = new Array(parts)
   let sum = 0
+  let hasNegativeValues = true
   do {
     for (let i = 0; i < parts; i++) {
       arr[i] = Math.random()
@@ -11,7 +12,8 @@ const splitIntoRandomValues = (total, parts) => {
     const scale = (total - parts) / sum
     arr = arr.map((val) => Math.min(total, Math.round(val * scale) + 1))
     sum = arr.reduce((acc, val) => acc + val, 0)
-  } while (sum - total)
+    hasNegativeValues = arr.some((v) => v === -1)
+  } while (sum - total && hasNegativeValues)
 
   return arr
 }