+++ test-app/mocks/mockClients/dashboards/definitions/examples/scorecards/index.js
@@ -1,7 +1,9 @@
 const { dietTotalsScoreCards } = require('./scorecards-diet-totals')
 const { dataQualityScoreCards } = require('./scorecards-data-quality')
+const { docsScoreCard } = require('./scorecard-docs-example')
 
 module.exports = {
   dietTotalsScoreCards,
   dataQualityScoreCards,
+  docsScoreCard,
 }